Sync new version

+++ components/categories/EditCategoriesItem.tsx
@@ -1,19 +1,20 @@
-import React, { useContext, useEffect, useState } from 'react'
-import { View, Text, StyleSheet, Pressable, TextInput } from 'react-native'
-import { Colors } from '../../constants/colors'
-import Button from '../../util-components/Button'
-import IconButton from '../../util-components/IconButton'
-import { AuthContext } from '../../store/auth-context'
-import { popupMessage } from '../../util-components/PopupMessage'
-import { CategoryTypes } from '../../types/allTsTypes'
-import DropdownList from '../../util-components/DropdownList'
 import Constants from 'expo-constants';
+import React, { useContext, useEffect, useState } from 'react';
+import { Pressable, StyleSheet, Text, TextInput, View } from 'react-native';
+import { Colors } from '../../constants/colors';
+import { AuthContext } from '../../store/auth-context';
+import { useUser } from '../../store/user-context';
+import { CategoryTypes } from '../../types/allTsTypes';
+import Button from '../../util-components/Button';
+import DropdownList from '../../util-components/DropdownList';
+import IconButton from '../../util-components/IconButton';
+import { popupMessage } from '../../util-components/PopupMessage';
 const backendURI = Constants.expoConfig?.extra?.backendURI;
 
 interface DropdownTypes {
-  _id: string | number
-  name: string
-  value: string
+  _id: string | number;
+  name: string;
+  value: string;
 }
 
 function EditCategoriesItem({ data }: { data: CategoryTypes }) {
@@ -22,33 +23,34 @@ function EditCategoriesItem({ data }: { data: CategoryTypes }) {
     name: '',
     stockType: '',
   });
-  const [newName, setNewName] = useState('')
+  const [newName, setNewName] = useState('');
   const [showEdit, setShowEdit] = useState<Boolean>(false);
   const authCtx = useContext(AuthContext);
+  const user = useUser();
   const [success, setSucces] = useState('');
   const [error, setError] = useState('');
   const [display, setDisplay] = useState(true);
-  const [stockType, setStockType] = useState<DropdownTypes | undefined>()
+  const [stockType, setStockType] = useState<DropdownTypes | undefined>();
   const [dropdownData, setDropdownData] = useState<DropdownTypes[]>([
-    {_id: 0, name: 'Veličina', value: 'Boja-Veličina-Količina'},
-    {_id: 1, name: 'Boja', value: 'Boja-Količina'},
+    { _id: 0, name: 'Veličina', value: 'Boja-Veličina-Količina' },
+    { _id: 1, name: 'Boja', value: 'Boja-Količina' },
   ]);
 
-  // Resets Error & Success 
-  function resetNotifications(){
+  // Resets Error & Success
+  function resetNotifications() {
     setSucces('');
     setError('');
   }
 
   // Toggler
-  function showEditCategoryrHandler(){
-    setNewName(data.name)
+  function showEditCategoryrHandler() {
+    setNewName(data.name);
     setShowEdit(!showEdit);
   }
 
   // On input text change
   function handleNameChange(newName: string) {
-    setNewName(newName)
+    setNewName(newName);
   }
 
   // Set default data to read from
@@ -56,38 +58,43 @@ function EditCategoriesItem({ data }: { data: CategoryTypes }) {
     setCategoryData(data);
     setNewName(data.name);
     // setStockType(data.stockType);
-  }, [data])
+  }, [data]);
 
   // Updates the current category name in the database
-  async function updateCategoryHandler(){
-    try{
+  async function updateCategoryHandler() {
+    if (!user?.permissions?.categories?.update)
+      return popupMessage('Nemate permisiju za ažuriranje kategorija.', 'danger');
+    try {
       resetNotifications();
       // if(newName.trim() === data.name){
       //   setShowEdit(false);
       //   return;
       // }
-      if(newName.trim() === ''){
+      if (newName.trim() === '') {
         setError('Kategorija mora imati ime!');
         popupMessage('Kategorija mora imati ime', 'danger');
         return;
       }
-      if(!stockType){
+      if (!stockType) {
         setError('Kategorija mora imati jedinicu lagera!');
         popupMessage('Kategorija mora imati jedinicu lagera', 'danger');
         return;
       }
-      const response = await fetch(`${backendURI || process.env.EXPO_PUBLIC_BACKEND_URI}/categories/${categoryData._id}`, {
-        method: 'PATCH',
-        headers: {
-          'Authorization': `Bearer ${authCtx.token}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({ 
-          id: categoryData._id,
-          name: newName,
-          stockType: stockType.value
-        })
-      })
+      const response = await fetch(
+        `${backendURI || process.env.EXPO_PUBLIC_BACKEND_URI}/categories/${categoryData._id}`,
+        {
+          method: 'PATCH',
+          headers: {
+            Authorization: `Bearer ${authCtx.token}`,
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify({
+            id: categoryData._id,
+            name: newName,
+            stockType: stockType.value,
+          }),
+        }
+      );
 
       if (!response.ok) {
         const parsedResponse = await response.json();
@@ -99,22 +106,28 @@ function EditCategoriesItem({ data }: { data: CategoryTypes }) {
       setSucces(parsedResponse.message);
       popupMessage(parsedResponse.message, 'success');
       setShowEdit(false);
-    } catch(error){
+    } catch (error) {
       console.error('Error updating the category:', error);
     }
   }
 
   // Deletes the category from the database
-  async function removeCategoryHandler(){
+  async function removeCategoryHandler() {
+    if (!user?.permissions?.categories?.delete) {
+      return popupMessage('Nemate permisiju za brisanje kategorija.', 'danger');
+    }
     setDisplay(false);
-    try{
-      const response = await fetch(`${backendURI || process.env.EXPO_PUBLIC_BACKEND_URI}/categories/${categoryData._id}`, {
-        method: 'DELETE',
-        headers: {
-          'Authorization': `Bearer ${authCtx.token}`,
-          'Content-Type': 'application/json',
-        },
-      })
+    try {
+      const response = await fetch(
+        `${backendURI || process.env.EXPO_PUBLIC_BACKEND_URI}/categories/${categoryData._id}`,
+        {
+          method: 'DELETE',
+          headers: {
+            Authorization: `Bearer ${authCtx.token}`,
+            'Content-Type': 'application/json',
+          },
+        }
+      );
 
       if (!response.ok) {
         const parsedResponse = await response.json();
@@ -125,19 +138,17 @@ function EditCategoriesItem({ data }: { data: CategoryTypes }) {
       }
 
       popupMessage(`Kategorija je uspesno obrisana`, 'success');
-    } catch(error){
+    } catch (error) {
       console.error('Error deleting category:', error);
     }
   }
 
-  if(display === false){
+  if (display === false) {
     return;
   }
 
-  if(categoryData === null){
-    return (
-      <></>
-    )
+  if (categoryData === null) {
+    return <></>;
   }
 
   // function getStockTypeName(){
@@ -146,15 +157,12 @@ function EditCategoriesItem({ data }: { data: CategoryTypes }) {
   //   return '';
   // }
   return (
-    <Pressable 
-      style={styles.colorItem}
-      onPress={showEditCategoryrHandler}
-    >
+    <Pressable style={styles.colorItem} onPress={showEditCategoryrHandler}>
       {showEdit ? (
         <View style={styles.mainInputsContainer}>
-          <TextInput 
+          <TextInput
             style={styles.input}
-            placeholder='Ime kategorije'
+            placeholder="Ime kategorije"
             value={newName}
             onChangeText={handleNameChange}
           />
@@ -165,34 +173,34 @@ function EditCategoriesItem({ data }: { data: CategoryTypes }) {
             buttonContainerStyles={styles.dropdown}
           />
           <View style={styles.buttons}>
-            <Button 
+            <Button
               containerStyles={styles.buttonStyle}
               onPress={showEditCategoryrHandler}
               textColor={Colors.primaryLight}
               backColor={Colors.error}
-            >Otkazi</Button>
-            <Button 
+            >
+              Otkaži
+            </Button>
+            <Button
               containerStyles={styles.buttonStyle}
               onPress={updateCategoryHandler}
               textColor={Colors.primaryLight}
               backColor={Colors.primaryDark}
-            >Sacuvaj</Button>
+            >
+              Sačuvaj
+            </Button>
           </View>
-          {error && (
-            <Text style={styles.error}>{error}</Text>
-          )}
-          {success && (
-            <Text style={styles.success}>{success}</Text>
-          )}
+          {error && <Text style={styles.error}>{error}</Text>}
+          {success && <Text style={styles.success}>{success}</Text>}
         </View>
       ) : (
         <View style={styles.displayCategory}>
           <View style={styles.categoryData}>
-              <Text style={[styles.text, styles.categoryName]}>{categoryData.name}</Text>
-              <Text style={styles.text}>{data.stockType}</Text>
+            <Text style={[styles.text, styles.categoryName]}>{categoryData.name}</Text>
+            <Text style={styles.text}>{data.stockType}</Text>
           </View>
           <IconButton
-            icon='delete'
+            icon="delete"
             onPress={removeCategoryHandler}
             color={Colors.error}
             style={styles.deleteIcon}
@@ -201,20 +209,21 @@ function EditCategoriesItem({ data }: { data: CategoryTypes }) {
         </View>
       )}
     </Pressable>
-  )
+  );
 }
 
 const styles = StyleSheet.create({
   colorItem: {
     padding: 14,
     paddingHorizontal: 25,
-    borderBottomWidth: 1,
-    borderBottomColor: '#ccc',
+    borderWidth: 0.5,
+    borderColor: Colors.secondaryLight,
     backgroundColor: 'white',
     marginBottom: 1,
     flexDirection: 'row',
     gap: 20,
-    alignItems: 'center'
+    alignItems: 'center',
+    elevation: 1,
   },
   deleteIcon: {
     marginLeft: 'auto',
@@ -227,7 +236,7 @@ const styles = StyleSheet.create({
   },
   categoryData: {
     flexDirection: 'column',
-    width: '100%'
+    width: '100%',
   },
   text: {
     fontSize: 14,
@@ -241,14 +250,14 @@ const styles = StyleSheet.create({
     flexDirection: 'column',
   },
   input: {
-    borderBottomColor: Colors.primaryDark,
+    borderBottomColor: Colors.secondaryLight,
     borderBottomWidth: 1,
     flex: 1,
-    marginBottom: 10, 
+    marginBottom: 10,
     fontSize: 16,
   },
   dropdown: {
-    marginBottom: 8
+    marginBottom: 8,
   },
   buttons: {
     flexDirection: 'row',
@@ -263,11 +272,11 @@ const styles = StyleSheet.create({
   success: {
     marginTop: 8,
     color: Colors.success,
-    textAlign: 'center'
+    textAlign: 'center',
   },
   buttonStyle: {
-    flex: 1
-  }
-})
+    flex: 1,
+  },
+});
 
-export default EditCategoriesItem
+export default EditCategoriesItem;
